projects/datanext/rootpanel/requirements.txt
playwright>=1.52.0
pytest>=8.3.5
python-dotenv>=1.0.1
